Quick minor changes, should behave the same: - When Nassim was integrating my part to the server, he created a new activity on top of what I already had, I just merged them together and added some methods for testing purposes. - Removed separate activities for leaving review for Rider/Driver, so now there is only ‘LeaveReviewFragment’.
