--- mRides-app/Resources/layout/DriverReview.axml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:minWidth="25px"
-    android:minHeight="25px">
-    <TextView
-        
-         android:text="@string/leaveReviewForUser"
-        android:textSize="20dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/driverReviewText"
-        android:padding="10dp" />
-    <RatingBar
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/ratingBarDriver"
-        android:numStars="5"
-        android:stepSize="0.1"
-        android:progressTint="#00ff00"
-        android:layout_gravity="center" />
-    <EditText
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="30dp"
-        android:id="@+id/driverReviewEdit"
-        android:textColor="#ffffff" />
-    <Button
-        android:text="@string/reviewSubmit"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/submitFeedback1" />
-</LinearLayout>
